requirements.txt
datasets==3.6.0
jiwer==3.1.0
matplotlib==3.10.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
paddleocr==2.8.1
scikit-image==0.24.0
Pillow==11.2.1
sacrebleu==2.5.1
simple_lama_inpainting
transformers==4.51.3
paddlepaddle==2.6.1
onnxruntime==1.18.1
fastapi==0.85.0
uvicorn[standard]==0.18.3
numpy==1.26.4
python-multipart==0.0.9
sentencepiece==0.2.0
fire==0.5.0